Rename gltf module to mesh, break up some files into submodules
