+++ demo/src/assets/mesh/assets/mesh_asset.rs
@@ -259,120 +259,3 @@ impl DefaultAssetTypeLoadHandler<MeshAssetData, MeshAsset> for MeshLoadHandler {
 }
 
 pub type MeshAssetType = DefaultAssetTypeHandler<MeshAssetData, MeshAsset, MeshLoadHandler>;
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct ModelAssetDataLod {
-    pub mesh: Handle<MeshAsset>,
-}
-
-#[derive(TypeUuid, Serialize, Deserialize, Clone)]
-#[uuid = "75bbc873-e527-42c6-8409-15aa5e68a4a4"]
-pub struct ModelAssetData {
-    pub lods: Vec<ModelAssetDataLod>,
-}
-
-pub struct ModelAssetInner {
-    pub lods: Vec<ModelAssetDataLod>,
-}
-
-#[derive(TypeUuid, Clone)]
-#[uuid = "76b953ef-9d1e-464b-b2a8-74f5b8842bd8"]
-pub struct ModelAsset {
-    pub inner: Arc<ModelAssetInner>,
-}
-
-pub struct ModelLoadHandler;
-
-impl DefaultAssetTypeLoadHandler<ModelAssetData, ModelAsset> for ModelLoadHandler {
-    #[profiling::function]
-    fn load(
-        _asset_manager: &mut AssetManager,
-        model_asset: ModelAssetData,
-    ) -> RafxResult<ModelAsset> {
-        let inner = ModelAssetInner {
-            lods: model_asset.lods,
-        };
-
-        Ok(ModelAsset {
-            inner: Arc::new(inner),
-        })
-    }
-}
-
-pub type ModelAssetType = DefaultAssetTypeHandler<ModelAssetData, ModelAsset, ModelLoadHandler>;
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct PrefabAssetDataObjectTransform {
-    pub position: glam::Vec3,
-    pub rotation: glam::Quat,
-    pub scale: glam::Vec3,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct PrefabAssetDataObjectModel {
-    pub model: Handle<ModelAsset>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Copy, Debug)]
-pub enum PrefabAssetDataObjectLightKind {
-    Point,
-    Spot,
-    Directional,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct PrefabAssetDataObjectLightSpot {
-    pub inner_angle: f32,
-    pub outer_angle: f32,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct PrefabAssetDataObjectLight {
-    pub color: glam::Vec3,
-    pub kind: PrefabAssetDataObjectLightKind,
-    pub intensity: f32,
-    pub spot: Option<PrefabAssetDataObjectLightSpot>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct PrefabAssetDataObject {
-    pub transform: PrefabAssetDataObjectTransform,
-    pub model: Option<PrefabAssetDataObjectModel>,
-    pub light: Option<PrefabAssetDataObjectLight>,
-}
-
-#[derive(TypeUuid, Serialize, Deserialize, Clone, Debug)]
-#[uuid = "1af63a91-de3e-48fc-8908-ab309730b8b5"]
-pub struct PrefabAssetData {
-    pub objects: Vec<PrefabAssetDataObject>,
-}
-
-pub struct PrefabAssetInner {
-    pub objects: Vec<PrefabAssetDataObject>,
-}
-
-#[derive(TypeUuid, Clone)]
-#[uuid = "7bf45a97-62f4-4a9a-99b8-e0ac8d755993"]
-pub struct PrefabAsset {
-    pub inner: Arc<PrefabAssetInner>,
-}
-
-pub struct PrefabLoadHandler;
-
-impl DefaultAssetTypeLoadHandler<PrefabAssetData, PrefabAsset> for PrefabLoadHandler {
-    #[profiling::function]
-    fn load(
-        _asset_manager: &mut AssetManager,
-        model_asset: PrefabAssetData,
-    ) -> RafxResult<PrefabAsset> {
-        let inner = PrefabAssetInner {
-            objects: model_asset.objects,
-        };
-
-        Ok(PrefabAsset {
-            inner: Arc::new(inner),
-        })
-    }
-}
-
-pub type PrefabAssetType = DefaultAssetTypeHandler<PrefabAssetData, PrefabAsset, PrefabLoadHandler>;
