+++ demo/src/assets/mesh/assets/mod.rs
@@ -1,0 +1,8 @@
+pub mod mesh_asset;
+pub use mesh_asset::*;
+
+pub mod model_asset;
+pub use model_asset::*;
+
+pub mod prefab_asset;
+pub use prefab_asset::*;
