+++ demo/src/assets/mesh/importers/blender_material_importer.rs
@@ -1,5 +1,5 @@
 use super::MeshMaterialData;
-use crate::assets::gltf::MeshMaterialDataShaderParam;
+use crate::assets::mesh::MeshMaterialDataShaderParam;
 use distill::importer::{ImportedAsset, Importer, ImporterValue};
 use distill::make_handle_from_str;
 use distill::{core::AssetUuid, importer::ImportOp};
