+++ demo/src/assets/mesh/mod.rs
@@ -1,0 +1,8 @@
+mod assets;
+pub use assets::*;
+
+mod importers;
+pub use importers::*;
+
+mod plugin;
+pub use plugin::*;
