+++ demo/src/assets/mod.rs
@@ -1,3 +1,3 @@
 pub mod font;
-pub mod gltf;
 pub mod ldtk;
+pub mod mesh;
