+++ demo/src/features/mesh/internal/frame_packet.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::assets::gltf::MeshAsset;
+use crate::assets::mesh::MeshAsset;
 use crate::components::{
     DirectionalLightComponent, PointLightComponent, SpotLightComponent, TransformComponent,
 };
